% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.r
\name{app_server}
\alias{app_server}
\title{Server side logic of the application}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{shiny input components}

\item{output}{shiny output components}

\item{session}{the shiny session parameter}
}
\value{
ignored
}
\description{
Server side logic of the application
}
