% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.r
\name{create_config}
\alias{create_config}
\title{Configuration}
\usage{
create_config(dir = ".", force = FALSE)
}
\arguments{
\item{dir}{Folder to put config file}

\item{force}{Overwrite current config file}
}
\description{
Creates a configuration file based on the default
shipped with helseatlas package.
}
