% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterOut.r
\name{filterOut}
\alias{filterOut}
\title{Filter routine}
\usage{
filterOut(datasett, filter1, filter2, filter3)
}
\arguments{
\item{datasett}{Original dataset}

\item{filter1}{level1 filter}

\item{filter2}{level2 filter}

\item{filter3}{level3 filter}
}
\value{
datasett
}
\description{
Filter routine
}
