% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.r
\name{launch_app}
\alias{launch_app}
\title{Launch the application locally}
\usage{
launch_app(datasett = NULL, language = NULL, title = NULL)
}
\arguments{
\item{datasett}{The data set to be loaded into the application}

\item{language}{The language of the atlas ("en" or "no")}

\item{title}{The title of the atlas}
}
\description{
Launch the application locally
}
