% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.r
\name{make_map}
\alias{make_map}
\title{Function to make a map}
\usage{
make_map(data = NULL, map = NULL, type = "leaflet", utm33 = FALSE)
}
\arguments{
\item{data}{Data to be plotted in the map}

\item{map}{The map itself}

\item{type}{Type of map (default = "leaflet")}

\item{utm33}{Convert from UTM 33 to Leaflet projection (default = TRUE)}
}
\value{
a map
}
\description{
Function to make a map
}
