% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLeafletmap.r
\name{makeLeafletmap}
\alias{makeLeafletmap}
\title{Make a Leaflet map}
\usage{
makeLeafletmap(inputData = NULL)
}
\arguments{
\item{inputData}{The input data set to be plotted}
}
\value{
a map
}
\description{
Make a Leaflet map
}
