% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.r
\name{makeMap}
\alias{makeMap}
\title{Common function to make a map}
\usage{
makeMap(data = NULL, map = NULL, type = "leaflet")
}
\arguments{
\item{data}{Data to be plotted in the map}

\item{map}{The map itself}

\item{type}{Type of map (default = "leaflet")}
}
\value{
a map
}
\description{
Common function to make a map
}
