% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natural_breaks.r
\name{natural_breaks}
\alias{natural_breaks}
\title{Get the jenks natural breaks from a data set}
\usage{
natural_breaks(data = NULL, num = 4)
}
\arguments{
\item{data}{Input data}

\item{num}{Number of groups}
}
\value{
Natural breaks
}
\description{
Get the jenks natural breaks from a data set
}
