% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.r
\name{plotLeafletmap}
\alias{plotLeafletmap}
\title{Make a Leaflet map}
\usage{
plotLeafletmap(data = NULL, map = NULL, utm33 = TRUE)
}
\arguments{
\item{data}{Data to be presented in map}

\item{map}{Map to be plotted}

\item{utm33}{Convert from UTM 33 to Leaflet projection}
}
\value{
a map
}
\description{
Make a Leaflet map
}
