% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.r
\name{plotSimpleMap}
\alias{plotSimpleMap}
\title{Plot a map with data}
\usage{
plotSimpleMap(data = NULL, map = NULL, utm33 = TRUE)
}
\arguments{
\item{data}{Data to be plotted}

\item{map}{Map to plot}

\item{utm33}{Convert from UTM 33 to Leaflet projection}
}
\value{
a plot made from a geojson map
}
\description{
Plot a map with data
}
