% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVariation.r
\name{plotVariation}
\alias{plotVariation}
\title{Make a plot based on a data frame}
\usage{
plotVariation(fullDataFrame = NULL, which_id = "i0")
}
\arguments{
\item{fullDataFrame}{Data to be plotted}

\item{which_id}{Which data to plot (with id == which_id)}
}
\description{
Make a plot based on a data frame
}
