% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp2geojson.R
\name{reduce_map_size}
\alias{reduce_map_size}
\title{Reduce size of map}
\usage{
reduce_map_size(map, amount = 0.1)
}
\arguments{
\item{map}{Map to be reduced in size}

\item{amount}{How much to reduce the size}
}
\value{
Map of reduced size
}
\description{
Reduce size of map
}
