% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp2geojson.R
\name{shp2geojson}
\alias{shp2geojson}
\title{Function to convert a shapefile to geojson}
\usage{
shp2geojson(shapefile = "eldre", folder = ".",
  geojson = "eldre_reduced", reduce_size = TRUE, amount = 0.1)
}
\arguments{
\item{shapefile}{The basename of shapefile to be converted to geojson}

\item{folder}{The folder where the shapefile is}

\item{geojson}{The name of the geojson file to be created}

\item{reduce_size}{Should the file size be reduced?}

\item{amount}{How much to reduce the file size}
}
\description{
Function to convert a shapefile to geojson
}
