% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skde_colors.r
\name{skde_colors}
\alias{skde_colors}
\title{Function that returns a list of SKDE colors.}
\usage{
skde_colors(num = 4)
}
\arguments{
\item{num}{Number of SKDE colors to be returned}
}
\value{
List of colors in HEX format
}
\description{
Function that returns a list of SKDE colors.
}
