% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.r
\name{submit_app}
\alias{submit_app}
\title{Submit the application to shinyapp.io}
\usage{
submit_app(datasett = NULL, name = "experimental", HNproxy = FALSE,
  shiny_account = "skde", language = NULL, title = NULL)
}
\arguments{
\item{datasett}{The data set file (.RData) to be loaded into the application.
The absolute path has to be given}

\item{name}{The appName of the deployed shiny application (default = "experimental")}

\item{HNproxy}{If TRUE: deploy app through proxy}

\item{shiny_account}{Which shiny account on shinyapps.io (default = "skde")}

\item{language}{The language of the atlas ("en" or "no")}

\item{title}{The title of the atlas}
}
\description{
Submit the application to shinyapp.io
}
