% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shp2geojson.R
\name{utm33_to_leaflet}
\alias{utm33_to_leaflet}
\title{Convert map from UTM 33 projection to epsg:4326 projection}
\usage{
utm33_to_leaflet(map)
}
\arguments{
\item{map}{map to be converted}
}
\value{
Converted map
}
\description{
Convert map from UTM 33 projection to epsg:4326 projection
}
