% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_info.r
\name{version_info}
\alias{version_info}
\title{Functions providing informational text about the app(s)}
\usage{
version_info(newline = "<br>")
}
\arguments{
\item{newline}{String element defining line break for formatting. Default is
\code{<br>}}
}
\value{
String formatted in some sensible manner
}
\description{
Functions providing informational text about the app(s)
}
\examples{
version_info()
}
