% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.r
\name{create_appdir}
\alias{create_appdir}
\title{Create an appDir for shiny::runApp and rsconnect::deployApp}
\usage{
create_appdir(healthatlas_data = NULL, healthatlas_map = NULL,
  webpage_title = NULL)
}
\arguments{
\item{healthatlas_data}{The data to be saved in the directory, to be used by the app}

\item{healthatlas_map}{The map}

\item{webpage_title}{The title of the atlas}
}
\value{
The created directory
}
\description{
Create a directory in tempdir() where the installed version of
shinymap package is copied and the data is saved.
This directory, with its content, will be deployd to or ran by shiny.
}
