% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.r
\name{launch_app}
\alias{launch_app}
\title{Launch the application, either locally or to shinyapps.io}
\usage{
launch_app(dataset = NULL, map = NULL, title = NULL,
  publish_app = FALSE, name = "experimental", shiny_account = "skde",
  git_hash = NULL, github_repo = NULL)
}
\arguments{
\item{dataset}{The data set to be loaded into the application}

\item{map}{The map to be used in the application}

\item{title}{The title of the atlas}

\item{publish_app}{If TRUE: deploy app to shinyapps.io (default = FALSE)}

\item{name}{The appName of the deployed shiny application (default = "experimental")}

\item{shiny_account}{Which shiny account on shinyapps.io (default = "skde")}

\item{git_hash}{Current git sha1 hash}

\item{github_repo}{Current github repository}
}
\description{
Launch the application, either locally or to shinyapps.io
}
