% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_application.R
\name{launch_application}
\alias{launch_application}
\title{Launch the application locally}
\usage{
launch_application(datasett = NULL)
}
\arguments{
\item{datasett}{The data set to be loaded into the application}
}
\description{
Launch the application locally
}
