% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.r
\name{make_map}
\alias{make_map}
\title{Function to make a map}
\usage{
make_map(data = NULL, map = NULL)
}
\arguments{
\item{data}{Data to be plotted in the map}

\item{map}{The map itself}
}
\value{
a map
}
\description{
Function to make a map
}
