% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_variation.r
\name{plotVariation}
\alias{plotVariation}
\title{Make a plot based on a data frame}
\usage{
plotVariation(inputData = NULL, xlab = "Opptaksomr", ylab = "Rate",
  type = "histogram")
}
\arguments{
\item{inputData}{Data to be plotted}

\item{xlab}{Label on x-axis}

\item{ylab}{Label on y-axis}

\item{type}{What kind of plot. Default and only choice, at the moment:
"histogram"}
}
\description{
Make a plot based on a data frame
}
