% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readJS.R
\name{readIAjson}
\alias{readIAjson}
\title{Read json-data from IA}
\usage{
readIAjson(json_file = NULL)
}
\arguments{
\item{json_file}{The json file used by IA}
}
\value{
A data frame
}
\description{
Read json-data from IA
}
